motor de optimización de rutas, corrección de auth y actualización de docs
- **Optimización de Rutas:** Implementación de algoritmo híbrido (Vecino Más Cercano + Recocido Simulado) para reordenar waypoints eficientemente. Se añadieron los servicios `geometry.go` y `solver.go`.
- **API:** Nuevo endpoint `POST /routes/:id/optimize` restringido a Admins y Super Admins.
- **Corrección de Errores:**
  - Resuelto error `EOF` en `/auth/register` manejando cuerpos vacíos para peticiones de login vs registro.
  - Corregido tipo de dato `TotalDistanceKm` (cambiado de int a float64) en `domains/route.go`.
  - Corregida lógica de punteros para `FleetCode` en `users/update.go`.
  - Secciones de Dashboard y Optimización añadidas.

+++ api/handlers/routes/update.go
@@ -12,7 +12,7 @@ import (
 type UpdateRouteInput struct {
 	Name                 string     `json:"name"`
 	ScheduledDate        *time.Time `json:"scheduled_date"`
-	TotalDistanceKm      int        `json:"total_distance_km"`
+	TotalDistanceKm      float64    `json:"total_distance_km"`
 	EstimatedDurationMin int        `json:"estimated_duration_min"`
 }
